Editor Engine Base, etc. (#7)
* Feat : Launch, EditorEngineLoop

* Add : Editor class

* Feat : Editor 기본 구조 생성

* For Merge : Commit remain files

* Add : Editor Resizing, etc.

* Feat: PlatformTime 추가

* For Merge : Commit remain files

* Add : Launch/EngineLoop

* Feat : EngineLoop Interface

---------

+++ Engine/Source/Core/HAL/PlatformTime.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include "HAL/PlatformTypes.h"
+#include "Core/HAL/PlatformTypes.h"
 
 #if defined(_WIN32)
 #include <Windows.h>
@@ -9,10 +9,9 @@
 #include <thread>
 #endif
 
-// TODO
 struct FPlatformTime
 {
-    static double Seconds();
+    static float Seconds();
     static uint64 Cycles64();
     static void   Sleep(float Seconds);
 };
